+++ Solution/RiskOfChaos/RiskOfChaosPlugin.cs
@@ -26,7 +26,7 @@ namespace RiskOfChaos
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "Gorakh";
         public const string PluginName = "RiskOfChaos";
-        public const string PluginVersion = "2.7.0";
+        public const string PluginVersion = "2.7.1";
 
         Harmony _harmonyInstance;